chore(react): improvements in index.d.ts
