+++ packages/botonic-react/src/index.d.ts
@@ -1,0 +1,72 @@
+import * as React from 'react'
+import * as core from '@botonic/core'
+import { MessageType } from './components'
+/**
+ * See @botonic/core's Response for the description of the Response's semantics*/
+export interface BotResponse extends core.BotRequest {
+  response: [React.ReactNode]
+}
+
+export class NodeApp {
+  constructor(options: core.BotOptions)
+
+  renderNode(args): string
+
+  input(request: core.BotRequest): BotResponse
+}
+
+// Parameters of the actions' botonicInit method
+export interface ActionRequest {
+  session: core.Session
+  params: { [key: string]: string }
+  input: core.Input
+  plugins: { [id: string]: core.Plugin }
+  defaultTyping: number
+  defaultDelay: number
+  lastRoutePath: string
+}
+
+export class BotonicInputTester {
+  constructor(app: NodeApp)
+
+  text(
+    inp: string,
+    session?: core.Session,
+    lastRoutePath?: string
+  ): Promise<string>
+
+  payload(
+    inp: string,
+    session?: core.Session,
+    lastRoutePath?: string
+  ): Promise<string>
+}
+
+export class BotonicOutputTester {
+  constructor(app: NodeApp)
+
+  text(out: string, replies?: any): Promise<string>
+}
+
+export interface RequestContextInterface extends ActionRequest {
+  getString: (stringId: string) => string
+  setLocale: (locale: string) => string
+}
+
+export const RequestContext: React.Context<RequestContextInterface>
+
+export interface CustomMessageType {
+  customTypeName: string
+}
+
+export function msgToBotonic(
+  msg: any,
+  customMessageTypes?: CustomMessageType[],
+): React.ReactNode
+
+export function msgsToBotonic(
+  msgs: any | any[],
+  customMessageTypes?: CustomMessageType[],
+): React.ReactNode
+
+export * from './components'
